+++ src/Order/OrderStatus.php
@@ -24,6 +24,7 @@ use MyCLabs\Enum\Enum;
  * @method static OrderStatus UNPAID()
  * @method static OrderStatus REFUNDED()
  * @method static OrderStatus CANCELED()
+ * @method static OrderStatus INCOMPLETED()
  */
 final class OrderStatus extends Enum
 {
@@ -38,4 +39,5 @@ final class OrderStatus extends Enum
     private const UNPAID = "UNPAID";
     private const REFUNDED = "REFUNDED";
     private const CANCELED = "CANCELED";
+    private const INCOMPLETED = "INCOMPLETED";
 }